+++ src/random/random.hpp
@@ -6,7 +6,7 @@
 
 float     randMarkov(glm::vec3 transition, float initial);
 int       randBernoulli(float p);
-double    randUniform(double min, double max);
+float     randUniform(float min, float max);
 glm::vec3 randGaussian(double mu, double sigma, double xmin, double xmax, double height, double zmin, double zmax);
 glm::vec3 randBimodale(double min, double max, double peak1, double peak2, double stdDev1, double stdDev2);
 float     randExponential(int min, int max);
